Minimal fix for Go 1.18 compatibility
Unfortunately it seems like Go 1.18 has removed the ability to do `import "."`
in non-`go.mod` projects.

A more full fix for this is probably to create a go.mod with the canonical
import path... but as that may also break imports that were pulling this
project from another path, I haven't done that in this commit.

Personally / the source of this problem for  me is that go.uber.org/yarpc
uses tchannel which uses this, which fails with:
```
go.uber.org/yarpc imports
	github.com/uber/tchannel-go tested by
	github.com/uber/tchannel-go.test imports
	github.com/streadway/quantile tested by
	github.com/streadway/quantile.test imports
	.: "." is relative, but relative import paths are not supported in module mode
```